Adjustment of the remap forms.py
the 'forms.py' were earlier copied from the energy2go server at AD.
This file contained the Django FormWizard setup.
For renooble on AD, the FormWizard will be replaced by a simple
form with the preview form function of Django forms. The earlier
geoCoding tool is not needed any more since we can use the jQuery app
for asyncronous geocoding on the site.